requirements.txt
streamlit==1.31.0
langchain==0.1.11
langchain-openai==0.0.8
langchain-community==0.0.27
langchain-core==0.1.31
chromadb==0.4.22
pydantic==2.6.3
python-dotenv==1.0.1
pypdf==4.0.2
openai==1.12.0
tiktoken==0.6.0
